Properly patch the heap size for Open Watcom
The memory layout for OW is different from Turbo C.
For Turbo, heap is below stack, but for OW it's the reverse.
Hence OW keeps the data segment much more minimal by default.
There is however a default sbrk increment of 8096 set by
_amblksiz. If we set that to 6K instead the heap will stay
at that size, so TSRs are loaded beyond that point
(for non-XMS-SWAP that is, with XMS-SWAP it's completely
different).


git-svn-id: http://svn.code.sf.net/p/freedos/svn/freecom/trunk@1775 6ac86273-5f31-0410-b378-82cca8765d1b